test: migrate diffing tests
