+++ packages/integration-karma/test/rendering/table-diffing/x/row/row.html
@@ -1,0 +1,3 @@
+<template>
+    <span>{row.id}</span>
+</template>
