+++ packages/integration-karma/test/rendering/table-diffing/x/row/row.js
@@ -1,0 +1,5 @@
+import { LightningElement, api } from 'lwc';
+
+export default class Row extends LightningElement {
+    @api row;
+}
