+++ packages/integration-karma/test/rendering/table-diffing/x/table/table.html
@@ -1,0 +1,9 @@
+<template>
+    <template for:each={rows} for:item="row">
+        <x-row
+            key={row.id}
+            row={row}
+        >
+        </x-row>
+    </template>
+</template>
