+++ packages/integration-karma/test/rendering/table-diffing/x/table/table.js
@@ -1,0 +1,5 @@
+import { LightningElement, api } from 'lwc';
+
+export default class Table extends LightningElement {
+    @api rows = [];
+}
