+++ packages/integration-karma/test/services/hooks/index.spec.js
